+++ locale/zh-cn/script.lua
@@ -139,7 +139,7 @@ DIAG_REDUNDANT_RETURN_VALUE_RANGE     =
 DIAG_MISSING_RETURN                   =
 '此处需要返回值。'
 DIAG_RETURN_TYPE_MISMATCH             =
-'第 {index} 个返回值的类型为 `{def}` ，但实际返回的是 `{ref}`。'
+'第 {index} 个返回值的类型为 `{def}` ，但实际返回的是 `{ref}`。\n{err}'
 DIAG_UNKNOWN_OPERATOR                 =
 '未知的运算符 `{}`。'
 DIAG_UNREACHABLE_CODE                 =